create basice structure for the project
