+++ src/App.js
@@ -1,9 +1,18 @@
-import Nav from './components/rockets/Nav';
+import { Route, Routes } from 'react-router-dom';
+import Navigation from './components/features/Navigation';
+import Rockets from './components/features/rockets';
+import Missions from './components/features/missions';
+import Profile from './components/features/profile';
 
 function App() {
   return (
     <div>
-      <Nav />
+      <Navigation />
+      <Routes>
+        <Route path="/" element={<Rockets />} />
+        <Route path="/one" element={<Missions />} />
+        <Route path="/two" element={<Profile />} />
+      </Routes>
     </div>
   );
 }
