+++ src/components/features/missions.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Missions() {
+  return (
+    <div>
+      missions
+    </div>
+  );
+}
+
+export default Missions;
