+++ src/components/features/profile.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Profile() {
+  return (
+    <div>
+      profile
+    </div>
+  );
+}
+
+export default Profile;
