+++ src/components/features/rockets.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Rockets() {
+  return (
+    <div>
+      Rockets
+    </div>
+  );
+}
+
+export default Rockets;
