+++ src/components/styles/Nav.css
@@ -1,3 +1,7 @@
+li {
+  list-style: none;
+}
+
 .header-navigation {
   display: flex;
   justify-content: space-between;
@@ -34,7 +38,7 @@
 }
 
 .navigation-item {
-  list-style: none;
+  text-decoration: none;
   color: blue;
 }
